Implemented movement, added PXPolygon

+++ assignment04/PXEllipse.java
@@ -6,7 +6,7 @@ public class PXEllipse extends PXObject {
     Color color;
     int xPos, yPos, width, height;
 
-    public PXEllipse(int width, int height, int xPos, int yPos, int moveMult, Color color) {
+    public PXEllipse(int width, int height, int xPos, int yPos, float moveMult, Color color) {
         this.width = width;
         this.height = height;
         this.xPos = xPos - (width / 2);
@@ -19,7 +19,7 @@ public class PXEllipse extends PXObject {
     public void paint(int mouseDX, int mouseDY, Graphics g) {
         xPos += mouseDX * moveMult;
         yPos += mouseDY * moveMult;
-        
+
         g.setColor(color);
         g.drawOval(xPos, yPos, width, height);
     }
